pg_quality_analyzer/requirements.txt
psycopg2-binary>=2.9.5
pandas>=1.5.0
numpy>=1.22.0
pyyaml>=6.0
matplotlib>=3.6.0
seaborn>=0.12.0
jinja2>=3.1.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
scikit-learn>=1.2.0
markdownify>=0.11.0
tqdm>=4.65.0
nltk>=3.8.0
